# docs/swagger.yaml
basePath: /
definitions:
  auth.ChallengeOutput:
    properties:
      challenge:
        type: string
    type: object
...
  common.RawDicConfig:
    properties:
      ftxTokenAddres:
        items:
          $ref: '#/definitions/common.fa'
        type: array
      isDisableAllSign:
        description: 禁用所有签名
        type: boolean
      isDisableFtxSign:
        description: 禁用ftx签名
        type: boolean
      safePrices:
        items:
          $ref: '#/definitions/common.sp'
        type: array
    type: object
  common.ResBody:
    properties:
      code:
        example: 200
        type: integer
      data:
        description: Data interface{} `json:"data"`
        type: object
      msg:
        example: ok
        type: string
      time:
        type: string
    type: object
  common.fa:
    properties:
      ftxName:
        type: string
      tokenAddre:
        type: string
    type: object
  common.sp:
    properties:
      comment:
        description: 备注　可选项，可输入名字
        type: string
      max:
        type: number
      min:
        type: number
      tokenAddre:
        type: string
    type: object
...
  sys.loginModel:
    properties:
      address:
        description: |-
          Username string `form:"username" json:"username" binding:"required"`
          Password string `form:"password" json:"password" binding:"required"`
        type: string
      signature:
        type: string
    required:
    - address
    - signature
    type: object
  sys.loginRes:
    properties:
      address:
        type: string
      isAdmin:
        type: boolean
    required:
    - address
    - isAdmin
    type: object
info:
  contact:
    email: wuxiaofei@rechaintech.com
    name: 伍晓飞
  description: oracle-api接口文档.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://test.com/index.html
  title: oracle-api
  version: "1.0"
paths:
...
  /pub/dic_config:
    get:
      consumes:
      - application/json
      description: 字典配制信息
      operationId: ConfigHandler
      parameters:
      - description: 配制json
        in: query
        name: is_using
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: config json
          schema:
            $ref: '#/definitions/common.RawDicConfig'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 获取字典配制信息
      tags:
      - default
    post:
      consumes:
      - application/json
      description: 修改字典配制信息
      operationId: ConfigHandler
      parameters:
      - description: 配制json
        in: body
        name: dic
        required: true
        schema:
          $ref: '#/definitions/common.RawDicConfig'
      produces:
      - application/json
      responses:
        "200":
          description: config json
          schema:
            $ref: '#/definitions/common.RawDicConfig'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 修改字典配制信息
      tags:
      - default
...
  /pub/login:
    post:
      consumes:
      - application/json
      description: 钱包登陆
      operationId: LoginHandler
      parameters:
      - description: 钱包登陆对象
        in: body
        name: logreq
        required: true
        schema:
          $ref: '#/definitions/sys.loginModel'
      produces:
      - application/json
      responses:
        "200":
          description: 登陆结果
          schema:
            $ref: '#/definitions/sys.loginRes'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 钱包登陆
      tags:
      - sys
  /pub/pre_login:
    get:
      consumes:
      - application/json
      description: 钱包登陆Challenge
      operationId: ChallengeHandler
      parameters:
      - default: 0x24C93Aaec52539a60240BCd2E972AB672D33eD79
        description: 钱包地址
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Challenge　code
          schema:
            $ref: '#/definitions/auth.ChallengeOutput'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 钱包登陆Challenge
      tags:
      - sys
...
swagger: "2.0"